Turn on -Wconversion and fix complaints
Reasonable advice from C++ Weekly with Jason Turner:
https://www.youtube.com/watch?v=R6_PFqOSa_c

Also, removed the long long modifier for printf, since we weren't using
it properly and were falling back to using long.

+++ kernel/uart.hpp
@@ -144,7 +144,7 @@ private:
         }
 
     private:
-        const u8 m_prev;
+        const u32 m_prev;
     };
 
     using ReadInterruptMask = InterruptMask<UART_IMSC_RXIM, UART_IMSC_RXIM>;
